@@ -17,17 +17,17 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/forms": "^7.2.0",
+    "@angular/animations": "^8.0.2",
+    "@angular/common": "^8.0.2",
+    "@angular/compiler": "^8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/forms": "^8.0.2",
     "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular/platform-browser": "^8.0.2",
+    "@angular/platform-browser-dynamic": "^8.0.2",
+    "@angular/router": "^8.0.2",
     "@ngx-translate/core": "^11.0.1",
-    "@angular/service-worker": "^7.2.0",
+    "@angular/service-worker": "^8.0.2",
     "@angular/cdk": "^7.2.1",
     "@angular/material": "^7.2.1",
     "@angular/material-moment-adapter": "^7.2.1",
@@ -62,16 +62,15 @@
     "rxjs-compat": "^6.3.3",
     "ts-helpers": "^1.1.2",
     "tslib": "^1.9.0",
-    "zone.js": "^0.8.27"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
     "@angular-devkit/core": "^7.2.1",
     "@angular-devkit/schematics": "^7.2.1",
     "@angular/cli": "^8.0.3",
-    "@angular/compiler": "^7.2.0",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.0",
+    "@angular/compiler-cli": "^8.0.2",
+    "@angular/language-service": "^8.0.2",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
     "@ngrx/schematics": "^7.0.0",
     "@ngx-rocket/scripts": "^3.0.1",
