@@ -10,7 +10,7 @@ import { filter } from 'rxjs/operators';
 })
 export class ShellComponent implements OnInit {
 
-  @ViewChild('sidenav') sidenav: MatSidenav;
+  @ViewChild('sidenav', {static: true}) sidenav: MatSidenav;
 
   constructor(private media: ObservableMedia) { }
 
